requirement.txt
asgiref==3.4.1
astroid==2.6.5
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
boto3==1.18.41
botocore==1.21.41
certifi==2021.5.30
charset-normalizer==2.0.4
coverage==5.5
distlib==0.3.2
dj-database-url==0.5.0
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-ckeditor==6.1.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-on-heroku==1.1.2
django-pandas==0.6.4
django-phonenumber-field==5.2.0
django-storages==1.11.1
django-twilio==0.13.1.post0
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
install==1.3.4
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
phonenumbers==8.12.28
platformdirs==2.2.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.9.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tox==3.24.1
twilio==6.62.1
urllib3==1.26.6
virtualenv==20.7.0
whitenoise==5.3.0
wrapt==1.12.1
